% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{acc_stress}
\alias{acc_stress}
\title{Relationship Between Parameter and Stress}
\usage{
acc_stress(s = 1, rel = "2 * s / 5")
}
\arguments{
\item{s}{stress.}

\item{rel}{the relationship between parameter and stress.}
}
\value{
a numeric.
}
\description{
Relationship Between Parameter and Stress
}
