% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cumsub}
\alias{cumsub}
\title{Cumulative Substract}
\usage{
cumsub(x)
}
\arguments{
\item{x}{a numeric object, or an object that can be coerced to one of these.}
}
\value{
Returns a vector whose elements are the cumulative substracts of the elements of the argument.
}
\description{
Cumulative Substract
}
