% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Plot stochastic degradation process}
\usage{
main(group = 6, t = 1:200, para = c(1, 1))
}
\arguments{
\item{group}{group of the degradation path.}

\item{t}{time.}

\item{para}{parameters of a certain model.}
}
\value{
Return a list containing a data of degradation path and a figure.
}
\description{
It is an example of create a R package. This function is used to
make a set of stochastic degradation process path. We hope to
make it easy to xxx.
Thank you for using it.
}
\examples{
main(group = 6,t = 1:200,para = c(1,1))
}
