% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{MLE}
\alias{MLE}
\title{Generate MLE method}
\usage{
MLE(
  par = c(3, 3),
  data = NULL,
  process = "Wiener",
  type = "classical",
  s = NULL,
  rel = NULL
)
}
\arguments{
\item{par}{parameters of a certain model.}

\item{data}{The number of groups of products}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}

\item{s}{stress.}

\item{rel}{relationship.}
}
\value{
Return a list containing simulated data at different time points for each group.
The output is a list.
The first part of the list is the amount of degradation of the simulated data,
and the second part is the increment of its degradation
}
\description{
This function is used to generate MLE.
}
\examples{
1
}
