% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reliability.R
\name{Reliability}
\alias{Reliability}
\title{Calculate Reliability function}
\usage{
Reliability(
  t = 100,
  threshold = 150,
  par = mle_par,
  process = "Wiener",
  type = "classical"
)
}
\arguments{
\item{t}{time.}

\item{threshold}{threshold of a degradation path.}

\item{par}{parameters of a certain model.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to calculate reliability function.
}
\examples{
dat <- sim_dat(group = 6, t = 1:200, para = c(2, 3), process = "Wiener")
mle_fit <- sta_infer(method = "MLE", process = "Wiener", type = "classical", data = dat[[1]])
Reliability(t = 50, threshold = 150, par = mle_fit, process = "Wiener", type = "classical")
}
