% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reliability_cowplot.R
\name{Reliability_cowplot}
\alias{Reliability_cowplot}
\title{Reliability_cowplot}
\usage{
Reliability_cowplot(R_time, sum_para, threshold, process, type)
}
\arguments{
\item{R_time}{R_time}

\item{sum_para}{sum_para}

\item{threshold}{a vector of thresholds for each group.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}
}
\value{
Return a plot of panel
}
\description{
This function is used to calculate Remaining Useful Life (RUL).
}
\examples{
1
}
