% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reliability_plot.R
\name{Reliability_plot}
\alias{Reliability_plot}
\title{Reliability_plot}
\usage{
Reliability_plot(
  R_time = 1:300,
  sum_para = fit,
  threshold = 150,
  process = "Wiener",
  type = "classical"
)
}
\arguments{
\item{R_time}{R_time}

\item{sum_para}{sum_para}

\item{threshold}{threshold}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to calculate Remaining Useful Life (RUL).
}
\examples{
dat = sim_dat(group = 6, t = 1:200, para = c(2,3),process = "Wiener")
mle_fit = sta_infer(method = "MLE", process = "Wiener",type = "classical",
data = dat[[1]])
Reliability_plot(R_time = 1:150,sum_para = mle_fit,threshold = 150,
process = "Wiener",type = "classical")
}
