% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUL.R
\name{RUL}
\alias{RUL}
\title{Remaining Useful Life}
\usage{
RUL(
  t = NULL,
  cur_time = NULL,
  threshold = NULL,
  data = NULL,
  par = NULL,
  process = "Wiener",
  type = "classical"
)
}
\arguments{
\item{t}{time.}

\item{cur_time}{current time.}

\item{threshold}{threshold of a degradation path.}

\item{data}{degradation data.}

\item{par}{parameters of a certain model.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to calculate Remaining Useful Life (RUL).
}
\examples{
1
}
