% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUL_3D_density.R
\name{RUL_3D_density}
\alias{RUL_3D_density}
\title{Remaining Useful Life}
\usage{
RUL_3D_density(
  fut_time,
  time_epoch,
  rul_den,
  threshold,
  real_RUL,
  zlim = c(0, 0.1),
  xlim = c(0, 100)
)
}
\arguments{
\item{fut_time}{fut_time}

\item{time_epoch}{time_epoch}

\item{rul_den}{ul_den}

\item{threshold}{threshold}

\item{real_RUL}{real_RUL}

\item{zlim}{zlim}

\item{xlim}{xlim
data data.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to calculate Remaining Useful Life (RUL).
}
\examples{
dat = sim_dat(group = 6, t = 1:200, para = c(2,3),process = "Wiener")
mle_fit <- sta_infer(method = "MLE", process = "Wiener", type = "classical", data = dat[[1]])
RUL_plot(
  fut_time = c(50, 55, 60, 65, 70, 75, 80), time_epoch = 1:100, threshold = 150, zlim = c(0, 0.05),
  xlim = c(0, 100), para = mle_fit[, 2], real_RUL = c(NA, NA, NA, NA, NA, NA) + 40
)
}
