% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUL_plot.R
\name{RUL_plot}
\alias{RUL_plot}
\title{Remaining Useful Life Ploting}
\usage{
RUL_plot(
  fut_time = c(50, 55, 60, 65, 70, 75, 80),
  group = 1,
  time_epoch = 1:100,
  threshold = 350,
  data = NULL,
  para = NULL,
  process = "Wiener",
  type = "classical",
  zlim = c(0, 0.01),
  xlim = c(0, 300),
  real_RUL = NULL
)
}
\arguments{
\item{fut_time}{time.}

\item{group}{group}

\item{time_epoch}{time_epoch}

\item{threshold}{threshold of a degradation path.}

\item{data}{degradation data.}

\item{para}{parameters of a certain model.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}

\item{zlim}{zlim.}

\item{xlim}{xlim.}

\item{real_RUL}{real_RUL.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to calculate Remaining Useful Life (RUL).
}
\examples{
dat <- sim_dat(group = 5, t = 1:200, para = c(2,3),
process = "Wiener",type = "classical")
# MLE
mle_fit = sta_infer(method = "MLE", process = "Wiener",
type = "classical", data = dat)
# RUL
RUL_plot(fut_time = c(50,55,60,65,70,75,80),time_epoch = 1:100,
threshold = 150,zlim = c(0,0.05),xlim = c(0,100),data = dat[[1]],
para = mle_fit[,2], group = 1, real_RUL=c(NA,NA,NA,NA,NA,NA)+40)
}
