% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Null.R
\docType{package}
\name{sdp}
\alias{sdp}
\title{sdp : package to learn how to develop R package}
\description{
It is an example for R package development.
The main function of it is to create a plot,
which is consist of several random walk chains.
}
\section{1 The purpose of package}{

Function main is used to ... The key function is function main ..........
}

\section{2 xxx}{

We hope this case can help you learn about the development of a R package.
}

\section{3 xxx}{

We hope this case can help you learn about the development of a R package.
}

\section{4 Conclusion}{

We hope this case can help you learn about the development of a R package.
}

