% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat.R
\name{sim_dat}
\alias{sim_dat}
\title{Generate simulated degradation data}
\usage{
sim_dat(group, t, para, process)
}
\arguments{
\item{group}{The number of groups of products}

\item{t}{time.}

\item{para}{parameters of a certain model.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to generate simulated degradation data.
}
\examples{
dat <- sim_dat(group = 6, t = 1:200, para = c(2, 3), process = "Wiener")
head(dat)
}
