% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{sta_infer}
\alias{sta_infer}
\title{Statistical inference}
\usage{
sta_infer(
  method = "MLE",
  process = "Wiener",
  type = "classical",
  data = dat,
  p = 0.975,
  par = c(1, 1),
  chains = 1,
  warmup = 1000,
  iter = 2000,
  cores = 1,
  s = NULL,
  rel = NULL
)
}
\arguments{
\item{method}{time.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}

\item{data}{degradation data.}

\item{p}{in default `p=0.975`.}

\item{par}{initial parameter.}

\item{chains}{in default `chains = 1`}

\item{warmup}{in default `warmup = 1000`.}

\item{iter}{in default `iter= 2000`.}

\item{cores}{cores = 1,}

\item{s}{stress.}

\item{rel}{relationship.}
}
\value{
Returns a data frame if method is MLE,
returns stan raw results if method is Bayes.
}
\description{
This function is used to statistical inference of model parameters.
The methods including maximum likelihood estimator and Bayesian inference.
}
