% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cumsub}
\alias{cumsub}
\title{Remaining Useful Life}
\usage{
cumsub(x)
}
\arguments{
\item{x}{data.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
Remaining Useful Life
}
