% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utility.R
\name{plot_path}
\alias{plot_path}
\title{Remaining Useful Life}
\usage{
plot_path(data = liion_new)
}
\arguments{
\item{data}{data.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to calculate Remaining Useful Life (RUL).
}
\examples{
dat <- sim_dat(group = 6, t = 1:200, para = c(2, 3), process = "Wiener")
plot_path(dat[[1]])
}
