% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path.R
\name{plot_path}
\alias{plot_path}
\title{Remaining Useful Life}
\usage{
plot_path(data = liion_new)
}
\arguments{
\item{data}{data.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to calculate Remaining Useful Life (RUL).
}
\examples{
dat <- sim_dat(group = 5, t = 1:200, para = c(2,3),
process = "Wiener",type = "classical",
s = NULL, rel = NULL)
plot_path(dat)
# add ggplot related functions.
plot_path(dat) +
theme_bw() +
theme(panel.grid = element_blank())
}
