% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reliability.R
\name{Reliability}
\alias{Reliability}
\title{Calculate Reliability function}
\usage{
Reliability(
  t = 100,
  threshold = 150,
  par = NULL,
  process = "Wiener",
  type = "classical",
  s = NULL,
  rel = NULL
)
}
\arguments{
\item{t}{time.}

\item{threshold}{threshold of a degradation path.}

\item{par}{parameters of a certain model.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}

\item{s}{stress.}

\item{rel}{relationship.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to calculate reliability function.
}
\examples{
reb <- Reliability(
  t = 40, threshold = 100, par = c(1, 2),
  process = "Wiener", type = "classical"
)
reb
}
