% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reliability_plot.R
\name{Reliability_plot}
\alias{Reliability_plot}
\title{Reliability_plot}
\usage{
Reliability_plot(
  R_time = 1:300,
  sum_para = fit,
  threshold = 150,
  process = "Wiener",
  type = "classical",
  rel = rel,
  s = s
)
}
\arguments{
\item{R_time}{R_time}

\item{sum_para}{sum_para}

\item{threshold}{a vector of thresholds for each group.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}

\item{rel}{relationship.}

\item{s}{stress.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to calculate Remaining Useful Life (RUL).
}
