% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reliability_plot.R
\name{Reliability_plot}
\alias{Reliability_plot}
\title{Reliability_plot}
\usage{
Reliability_plot(
  R_time = 1:300,
  sum_para = fit,
  threshold = 150,
  process = "Wiener",
  type = "classical",
  rel = NULL,
  s = NULL
)
}
\arguments{
\item{R_time}{R_time}

\item{sum_para}{sum_para}

\item{threshold}{a vector of thresholds for each group.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}

\item{rel}{relationship.}

\item{s}{stress.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to calculate Remaining Useful Life (RUL).
}
\examples{
dat <- sim_dat(
  group = 5, t = 1:200, para = c(2, 3),
  process = "Wiener", type = "classical"
)
# MLE
mle_fit <- sta_infer(
  method = "MLE", process = "Wiener",
  type = "classical", data = dat
)
# Reliability
Reliability_plot(
  R_time = 1:200, sum_para = mle_fit, threshold = 100,
  process = "Wiener", type = "classical"
)
}
