% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUL_3D_density.R
\name{RUL_3D_density}
\alias{RUL_3D_density}
\title{Remaining Useful Life}
\usage{
RUL_3D_density(
  fut_time,
  time_epoch,
  rul_den,
  threshold,
  real_RUL,
  zlim = c(0, 0.1),
  xlim = c(0, 100)
)
}
\arguments{
\item{fut_time}{fut_time}

\item{time_epoch}{time_epoch}

\item{rul_den}{ul_den}

\item{threshold}{threshold}

\item{real_RUL}{real_RUL}

\item{zlim}{zlim}

\item{xlim}{xlim
data data.}
}
\value{
Return a list containing RUL at different time points for each group.
}
\description{
This function is used to calculate Remaining Useful Life (RUL).
}
