% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat.R
\name{sim_dat}
\alias{sim_dat}
\title{Generate simulated degradation data}
\usage{
sim_dat(
  group = 5,
  t = 1:100,
  para = c(2, 3),
  process = "Wiener",
  type = "classical",
  s = NULL,
  rel = NULL
)
}
\arguments{
\item{group}{The number of groups of products}

\item{t}{time.}

\item{para}{parameters of a certain model.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}

\item{s}{stress.}

\item{rel}{relationship.}
}
\value{
Return a list containing simulated data at different time points for each group.
The output is a list.
The first part of the list is the amount of degradation of the simulated data,
and the second part is the increment of its degradation
}
\description{
This function is used to generate simulated degradation data.
}
\examples{
dat <- sim_dat(group = 5, t = 1:200, para = c(2,3),
process = "Wiener",type = "classical",
s = NULL, rel = NULL)
str(dat)
}
