% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{sta_infer}
\alias{sta_infer}
\title{Statistical inference}
\usage{
sta_infer(method, process, type, data)
}
\arguments{
\item{method}{time.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}

\item{data}{degradation data.}
}
\value{
Returns a data frame if method is MLE,
returns stan raw results if method is Bayes.
}
\description{
This function is used to statistical inference of model parameters.
The methods including maximum likelihood estimator and Bayesian inference.
}
\examples{
11
}
