% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta_infer.R
\name{sta_infer}
\alias{sta_infer}
\title{Statistical inference}
\usage{
sta_infer(
  method = "MLE",
  process = "Wiener",
  type = "classical",
  data = NULL,
  p = 0.975,
  par = c(1, 1),
  chains = 1,
  warmup = 1000,
  iter = 2000,
  cores = 1,
  s = NULL,
  rel = NULL
)
}
\arguments{
\item{method}{time.}

\item{process}{Wiener, Gamma or Inverse Gaussian process.}

\item{type}{classical in default.}

\item{data}{degradation data.}

\item{p}{in default `p=0.975`.}

\item{par}{initial parameter.}

\item{chains}{in default `chains = 1`}

\item{warmup}{in default `warmup = 1000`.}

\item{iter}{in default `iter= 2000`.}

\item{cores}{cores = 1,}

\item{s}{stress.}

\item{rel}{relationship.}
}
\value{
Returns a data frame if method is MLE,
returns stan raw results if method is Bayes.
}
\description{
This function is used to statistical inference of model parameters.
The methods including maximum likelihood estimator and Bayesian inference.
}
\examples{
dat <- sim_dat(group = 5, t = 1:200, para = c(2,3),
process = "Wiener",type = "classical",
s = NULL, rel = NULL)
# MLE
mle_fit = sta_infer(method = "MLE", process = "Wiener",
type = "classical", data = dat)
mle_fit
# Bayes
# rstan_options(auto_write = TRUE)
# options(mc.cores = parallel::detectCores())
# bayes_fit = sta_infer(method = "Bayes", process = "Wiener",
# type = "classical", data = dat)
# bayes_fit$summary
# print(bayes_fit$summary, probs = c(0.025,0.5,0.975),pars = c("mu","w"))
# plot(bayes_fit$stan_re)
# traceplot(bayes_fit$stan_re,pars = c("mu","w"),
#            inc_warmup = TRUE,nrow = 1) +
#    theme(legend.position = "top")
}
